refactor(editor): studio thumbnail preview 분리
